Fredboat-566 Revert "Make AudioTrackContext inherit from NullableContext"
